+++ old_docs/make_headers.py
@@ -1,18 +1,18 @@
 # Read file
-with open('supporting_platforms.md') as f:
+with open("supporting_platforms.md") as f:
     lines = f.readlines()
 
 # Generate toc
-toc_string = ''
+toc_string = ""
 
 for line in lines:
-    if '### ' in line:
-        raw_header = line.removeprefix('### ')
+    if "### " in line:
+        raw_header = line.removeprefix("### ")
 
         # Replace special characters in link strings.
-        link_string = raw_header.replace(' ', '-')
-        for c in ['`', '*', '.']:
-            link_string = link_string.replace(c, '')
+        link_string = raw_header.replace(" ", "-")
+        for c in ["`", "*", "."]:
+            link_string = link_string.replace(c, "")
 
         link_string = link_string.lower().rstrip()
 
@@ -21,5 +21,5 @@ for line in lines:
         toc_string += toc_line
 
 # Write toc to tmp file.
-with open('tmp_header.md', 'w') as f:
+with open("tmp_header.md", "w") as f:
     f.write(toc_string)