+++ sample_project/blog_project/blog/urls.py
@@ -17,7 +17,7 @@ from django.contrib import admin
 from django.urls import path, include
 
 urlpatterns = [
-    path('admin/', admin.site.urls),
-    path('users/', include('users.urls')),
-    path('', include('blogs.urls')),
+    path("admin/", admin.site.urls),
+    path("users/", include("users.urls")),
+    path("", include("blogs.urls")),
 ]