+++ simple_deploy/management/commands/fly_io/templates/settings.py
@@ -1,6 +1,5 @@
 {{current_settings}}
 
-
 # Fly.io settings.
 import os