+++ simple_deploy/management/commands/heroku/tests/e2e_tests/utils.py
@@ -2,13 +2,14 @@
 
 import re, time
 
-from ...utils.it_helper_functions import make_sp_call
+from tests.e2e_tests.utils.it_helper_functions import make_sp_call
 
 
 def create_project():
     """Create a project on Heroku."""
     print("\n\nCreating a project on Heroku...")
     make_sp_call("heroku create")
+
 
 def push_project():
     """Make a non-automated deployment."""
@@ -25,22 +26,25 @@ def push_project():
     print("  Opening project...")
     make_sp_call("heroku open")
 
-def get_project_url_name():
-    """Get project URL and app name of the deployed project.
-    """
-    output = make_sp_call("heroku apps:info", capture_output=True).stdout.decode().strip()
 
-    re_app_name = r'=== (.*)'
+def get_project_url_name():
+    """Get project URL and app name of the deployed project."""
+    output = (
+        make_sp_call("heroku apps:info", capture_output=True).stdout.decode().strip()
+    )
+
+    re_app_name = r"=== (.*)"
     app_name = re.search(re_app_name, output).group(1).strip()
 
     print(f"  Found app name: {app_name}")
 
     # Get URL.
-    re_url = r'Web URL:.*(https:.*)'
+    re_url = r"Web URL:.*(https:.*)"
     project_url = re.search(re_url, output).group(1).strip()
     print(f"  Project URL: {project_url}")
 
     return project_url, app_name
+
 
 def destroy_project(request):
     """Destroy the deployed project, and all remote resources."""